проверка на NULL

+++ src/main/java/ru/perfumess/controllers/rest/v1/admin/ProductAdminController.java
@@ -1,6 +1,7 @@
 package ru.perfumess.controllers.rest.v1.admin;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -13,6 +14,7 @@ import ru.perfumess.services.ProductService;
 
 import javax.validation.Valid;
 
+@Slf4j
 @RestController
 @RequiredArgsConstructor
 @RequestMapping(path = "/api/v1/admin/products")
@@ -27,6 +29,10 @@ public class ProductAdminController {
             @RequestParam("brandId") Brand brand,
             @RequestParam("photoId") Photo photo
     ) {
+        if (brand == null || photo == null) {
+            log.info("[newProduct] Brand or Photo NOT FOUND");
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         Product newProduct = productMapper.toProduct(productDto);
         newProduct.addBrand(brand);
         newProduct.addPhoto(photo);
@@ -39,6 +45,10 @@ public class ProductAdminController {
             @PathVariable("id") Product product,
             @PathVariable("brandId") Brand brand
     ) {
+        if (product == null || brand == null) {
+            log.info("[setBrand] Product or Brand NOT FOUND");
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         product.setBrand(brand);
         Product updatedProduct = productService.save(product);
         return new ResponseEntity<>(productMapper.toDto(updatedProduct), HttpStatus.OK);
@@ -49,6 +59,10 @@ public class ProductAdminController {
             @PathVariable("id") Product product,
             @PathVariable("photoId") Photo photo
     ) {
+        if (product == null || photo == null) {
+            log.info("[setBrand] Product or Photo NOT FOUND");
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         product.addPhoto(photo);
         product = productService.save(product);
         return new ResponseEntity<>(productMapper.toDto(product), HttpStatus.OK);
@@ -59,6 +73,10 @@ public class ProductAdminController {
             @PathVariable("id") Product product,
             @PathVariable("photoId") Photo photo
     ) {
+        if (product == null || photo == null) {
+            log.info("[setBrand] Product or Photo NOT FOUND");
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         product.deletePhoto(photo);
         product = productService.save(product);
         return new ResponseEntity<>(productMapper.toDto(product), HttpStatus.OK);
